+++ webpack/sequences/step_tiles/__tests__/tile_read_pin_test.tsx
@@ -1,10 +1,11 @@
 import * as React from "react";
-import { TileReadPin } from "../tile_read_pin";
+import { TileReadPin, changeStep } from "../tile_read_pin";
 import { mount } from "enzyme";
 import { fakeSequence } from "../../../__test_support__/fake_state/resources";
 import { ReadPin } from "farmbot/dist";
 import { emptyState } from "../../../resources/reducer";
-
+const s = changeStep;
+debugger;
 describe("<TileReadPin/>", () => {
   function bootstrapTest() {
     const currentStep: ReadPin = {
