+++ webpack/sequences/step_tiles/tile_read_pin.tsx
@@ -10,28 +10,6 @@ import { editStep } from "../../api/crud";
 import { SequenceBodyItem, ReadPeripheral } from "farmbot";
 import { TaggedSequence } from "../../resources/tagged_resources";
 
-export const EMPTY_READ_PERIPHERAL: ReadPeripheral = {
-  kind: "read_peripheral",
-  args: { peripheral_id: 0 }
-};
-
-export const changeStep =
-  (replacement: SequenceBodyItem) =>
-    (step: Readonly<SequenceBodyItem>,
-      sequence: Readonly<TaggedSequence>,
-      index: number) => {
-      return editStep({
-        step,
-        sequence,
-        index,
-        executor(c) {
-          c.kind = replacement.kind;
-          c.args = replacement.args;
-          c.body = replacement.body;
-        }
-      });
-    };
-
 const convertToReadPeripheral = changeStep(EMPTY_READ_PERIPHERAL);
 
 export function TileReadPin(props: StepParams) {
